requirements.txt
pyyaml
discord.py
aiowrap
